YUK-16302 - MAC ID and User Id should NOT be editable for Honeywell WiFi thermostats

+++ yukon-client/common/src/com/cannontech/stars/dr/honeywell/HoneywellBuilder.java
@@ -117,7 +117,7 @@ public class HoneywellBuilder implements HardwareTypeExtensionProvider {
         Integer existingDeviceVendorUserId =
             honeywellWifiThermostatDao.getHoneywellWifiThermostat(deviceId).getDeviceVendorUserId();
         
-        if (existingDeviceVendorUserId != deviceVendorUserId) {
+        if (existingDeviceVendorUserId.intValue() != deviceVendorUserId.intValue()) {
             throw new DeviceVendorUserIdNotUpdatableException();
         }
